Move test support classes to own package
