+++ lightblue/src/test/java/org/esbtools/eventhandler/lightblue/testing/EmptyNotificationFactory.java
@@ -16,9 +16,11 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.testing;
 
 import org.esbtools.eventhandler.Notification;
+import org.esbtools.eventhandler.lightblue.LightblueRequester;
+import org.esbtools.eventhandler.lightblue.NotificationFactory;
 import org.esbtools.lightbluenotificationhook.NotificationEntity;
 
 import java.util.NoSuchElementException;
