+++ lightblue/src/test/java/org/esbtools/eventhandler/lightblue/testing/LightblueClientConfigurations.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.testing;
 
 import com.redhat.lightblue.client.LightblueClientConfiguration;
 import com.redhat.lightblue.client.integration.test.LightblueExternalResource;
